- Rolls back exposing getMedia details from the editor package - Refactors page details and the useSelect callbacks - Rolls back the generic data list components since we don't need them anywhere else (premature optimization)
