--- packages/edit-site/src/components/sidebar-navigation-data-list/style.scss
@@ -1,12 +1,0 @@
-
-
-.edit-site-sidebar-navigation_data-list__item {
-	.edit-site-sidebar-navigation_data-list__label {
-		color: $gray-600;
-		width: 100px;
-	}
-
-	.edit-site-sidebar-navigation_data-list__value.edit-site-sidebar-navigation_data-list__value {
-		color: $gray-200;
-	}
-}
