+++ packages/editor/src/private-apis.js
@@ -4,11 +4,9 @@
 import { ExperimentalEditorProvider } from './components/provider';
 import { lock } from './lockUnlock';
 import { EntitiesSavedStatesExtensible } from './components/entities-saved-states';
-import getFeaturedMediaDetails from './components/post-featured-image/get-featured-media-details';
 
 export const privateApis = {};
 lock( privateApis, {
 	ExperimentalEditorProvider,
 	EntitiesSavedStatesExtensible,
-	getFeaturedMediaDetails,
 } );
